requirements.txt
jq>=1,<2
mcp>=1,<2
python-dotenv>=1,<2
pytest>=8,<9
pytest-asyncio>=0.23,<1
pytest-socket>=0.7,<1
